src/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
certifi==2019.11.28
chardet==3.0.4
client==0.0.1
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-model-utils==4.0.0
django-quiz-app==0.5.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
get==2019.4.13
idna==2.9
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
mysql==0.0.2
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
packaging==20.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
pluggy==0.13.1
post==2019.4.13
prometheus-client==0.7.1
prompt-toolkit==3.0.4
public==2019.4.13
py==1.8.1
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
tox==3.14.6
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
